+++ CHANGELOG.md
@@ -29,7 +29,8 @@ x.x.x Release notes (yyyy-MM-dd)
 
 ### Enhancements
 
-* None.
+* Improve the exception message when comparing a linked property to another
+  property (unsupported).
 
 ### Bugfixes